tests/android/test_wikipedia.py
import allure
from appium.webdriver.common.appiumby import AppiumBy
from selene import have
from selene.support.shared import browser
from allure import step

from mobile_tests_lesson_13.model import app

@allure.step("Открываем страницу википедии 1")
def test_check_for_search_resuls_1():
    app.given_opened()

    allure.dynamic.tag("web тест проверка  ")
    print("ok ok ok")
    allure.dynamic.tag("web тест проверка  ")
    print("ok ok ok")

    with step('Search for content'):
        browser.element((AppiumBy.ACCESSIBILITY_ID, 'Search Wikipedia')).click()
        browser.element((AppiumBy.ID, 'org.wikipedia.alpha:id/search_src_text')).type('BrowserStack')

    with step('Content should be found'):
        browser.all(
            (AppiumBy.ID, 'org.wikipedia.alpha:id/page_list_item_title')).should(have.size_greater_than(0))

@allure.step("Открываем страницу википедии 2")
def test_check_for_search_resuls_2():

    with step('Search for content'):
        browser.element((AppiumBy.ACCESSIBILITY_ID, 'Search Wikipedia')).click()
        browser.element((AppiumBy.ID, 'org.wikipedia.alpha:id/search_src_text')).type('yandex')

    with step('Content should be found'):
        browser.all(
            (AppiumBy.ID, 'org.wikipedia.alpha:id/page_list_item_title')).should(have.size_greater_than(0))
