requirements.txt
selenium ==4.5.0
selene==2.0.0b17
Appium-Python-Client == 2.8.1
pytest == 7.3.2
pytest-xdist ==2.5.0
allure-pytest == 2.10.0
requests == 2.28.1
pydantic == 1.10.5
python-dotenv==1.0.0